+++ jason/tests/api/room.rs
@@ -1,0 +1,176 @@
+#![cfg(target_arch = "wasm32")]
+
+use std::rc::Rc;
+
+use futures::{
+    sync::mpsc::{unbounded, UnboundedReceiver, UnboundedSender},
+    Future,
+};
+use medea_client_api_proto::{Event, IceServer, PeerId};
+use medea_jason::{
+    api::Room,
+    media::MediaManager,
+    peer::{MockPeerRepository, PeerConnection, PeerEvent},
+    rpc::MockRpcClient,
+};
+use wasm_bindgen::prelude::*;
+use wasm_bindgen_test::*;
+
+use crate::{get_test_tracks, resolve_after};
+
+wasm_bindgen_test_configure!(run_in_browser);
+
+fn get_test_room_and_exist_peer() -> (Room, Rc<PeerConnection>) {
+    let mut rpc = MockRpcClient::new();
+    let mut repo = Box::new(MockPeerRepository::new());
+    let (tx, _rx) = unbounded();
+    let peer = Rc::new(
+        PeerConnection::new(
+            PeerId(1),
+            tx,
+            vec![],
+            Rc::new(MediaManager::default()),
+            true,
+            true,
+        )
+        .unwrap(),
+    );
+
+    let (_, event_rx) = unbounded();
+    let peer_clone = Rc::clone(&peer);
+    rpc.expect_subscribe()
+        .return_once(move || Box::new(event_rx));
+    repo.expect_get_all()
+        .times(2)
+        .returning_st(move || vec![Rc::clone(&peer_clone)]);
+    rpc.expect_unsub().return_const(());
+
+    let room = Room::new(Rc::new(rpc), repo);
+    (room, peer)
+}
+
+#[wasm_bindgen_test(async)]
+fn mute_unmute_audio() -> impl Future<Item = (), Error = JsValue> {
+    let (room, peer) = get_test_room_and_exist_peer();
+    let (audio_track, video_track) = get_test_tracks();
+    peer.get_offer(vec![audio_track, video_track])
+        .map(move |_| {
+            let handle = room.new_handle();
+            assert!(handle.mute_audio().is_ok());
+            assert!(!peer.is_send_audio_enabled());
+            assert!(handle.unmute_audio().is_ok());
+            assert!(peer.is_send_audio_enabled());
+        })
+        .map_err(Into::into)
+}
+
+#[wasm_bindgen_test(async)]
+fn mute_unmute_video() -> impl Future<Item = (), Error = JsValue> {
+    let (room, peer) = get_test_room_and_exist_peer();
+    let (audio_track, video_track) = get_test_tracks();
+    peer.get_offer(vec![audio_track, video_track])
+        .map(move |_| {
+            let handle = room.new_handle();
+            assert!(handle.mute_video().is_ok());
+            assert!(!peer.is_send_video_enabled());
+            assert!(handle.unmute_video().is_ok());
+            assert!(peer.is_send_video_enabled());
+        })
+        .map_err(Into::into)
+}
+
+fn get_test_room_and_new_peer(
+    event_rx: UnboundedReceiver<Event>,
+    with_enabled_audio: bool,
+    with_enabled_video: bool,
+) -> (Room, Rc<PeerConnection>) {
+    let mut rpc = MockRpcClient::new();
+    let mut repo = Box::new(MockPeerRepository::new());
+
+    rpc.expect_subscribe()
+        .return_once(move || Box::new(event_rx));
+    repo.expect_get_all().returning(|| vec![]);
+    let (tx, _rx) = unbounded();
+    let peer = Rc::new(
+        PeerConnection::new(
+            PeerId(1),
+            tx,
+            vec![],
+            Rc::new(MediaManager::default()),
+            with_enabled_audio,
+            with_enabled_video,
+        )
+        .unwrap(),
+    );
+    let peer_clone = Rc::clone(&peer);
+    repo.expect_create_peer()
+        .withf(
+            move |id: &PeerId,
+                  _ice_servers: &Vec<IceServer>,
+                  _peer_events_sender: &UnboundedSender<PeerEvent>,
+                  enabled_audio: &bool,
+                  enabled_video: &bool| {
+                *id == PeerId(1)
+                    && *enabled_audio == with_enabled_audio
+                    && *enabled_video == with_enabled_video
+            },
+        )
+        .return_once_st(move |_, _, _, _, _| Ok(peer_clone));
+    rpc.expect_send_command().return_const(());
+    rpc.expect_unsub().return_const(());
+
+    let room = Room::new(Rc::new(rpc), repo);
+    (room, peer)
+}
+
+#[wasm_bindgen_test(async)]
+fn mute_audio_room_before_init_peer() -> impl Future<Item = (), Error = JsValue>
+{
+    let (event_tx, event_rx) = unbounded();
+    let (room, peer) = get_test_room_and_new_peer(event_rx, false, true);
+    let (audio_track, video_track) = get_test_tracks();
+
+    room.new_handle().mute_audio().unwrap();
+    event_tx
+        .unbounded_send(Event::PeerCreated {
+            peer_id: PeerId(1),
+            sdp_offer: None,
+            tracks: vec![audio_track, video_track],
+            ice_servers: vec![],
+        })
+        .unwrap();
+
+    resolve_after(500).and_then(move |_| {
+        // move room so it wont get dropped
+        let _ = room;
+        assert!(peer.is_send_video_enabled());
+        assert!(!peer.is_send_audio_enabled());
+        Ok(())
+    })
+}
+
+#[wasm_bindgen_test(async)]
+fn mute_video_room_before_init_peer() -> impl Future<Item = (), Error = JsValue>
+{
+    let (event_tx, event_rx) = unbounded();
+    let (room, peer) = get_test_room_and_new_peer(event_rx, true, false);
+    let (audio_track, video_track) = get_test_tracks();
+
+    room.new_handle().mute_video().unwrap();
+    event_tx
+        .unbounded_send(Event::PeerCreated {
+            peer_id: PeerId(1),
+            sdp_offer: None,
+            tracks: vec![audio_track, video_track],
+            ice_servers: vec![],
+        })
+        .unwrap();
+
+    resolve_after(500).and_then(move |_| {
+        // move room so it wont get dropped
+        let _ = room;
+        assert!(peer.is_send_audio_enabled());
+        assert!(!peer.is_send_video_enabled());
+        Ok(())
+    })
+}